+++ docs/_static/examples/gan.py
@@ -12,6 +12,8 @@ from torchvision.utils import save_image
 import torchbearer as tb
 import torchbearer.callbacks as callbacks
 from torchbearer import state_key
+from torchbearer.bases import base_closure
+
 
 os.makedirs('images', exist_ok=True)
 
@@ -37,10 +39,11 @@ G_LOSS = state_key('g_loss')
 D_LOSS = state_key('d_loss')
 
 DISC_OPT = state_key('disc_opt')
+GEN_OPT = state_key('gen_opt')
 DISC_MODEL = state_key('disc_model')
 DISC_IMGS = state_key('disc_imgs')
 DISC_CRIT = state_key('disc_crit')
-
+batch = torch.randn(25, latent_dim).to(device)
 
 
 class Generator(nn.Module):
@@ -60,7 +63,7 @@ class Generator(nn.Module):
             *block(256, 512),
             *block(512, 1024),
             nn.Linear(1024, int(np.prod(img_shape))),
-            nn.Tanh()
+            nn.Tanh(),
         )
 
     def forward(self, real_imgs, state):
@@ -89,10 +92,12 @@ class Discriminator(nn.Module):
 
         return validity
 
+
 def gen_crit(state):
     loss =  adversarial_loss(state[DISC_MODEL](state[tb.Y_PRED], state), valid)
     state[G_LOSS] = loss
     return loss
+
 
 def disc_crit(state):
     real_loss = adversarial_loss(state[DISC_MODEL](state[tb.X], state), valid)
@@ -101,13 +106,12 @@ def disc_crit(state):
     state[D_LOSS] = loss
     return loss
 
-batch = torch.randn(25, latent_dim).to(device)
+
 @callbacks.on_step_training
+@callbacks.only_if(lambda state: state[tb.BATCH] % sample_interval == 0)
 def saver_callback(state):
-    batches_done = state[tb.EPOCH] * len(state[tb.GENERATOR]) + state[tb.BATCH]
-    if batches_done % sample_interval == 0:
-        samples = state[tb.MODEL](batch, state)
-        save_image(samples, 'images/%d.png' % batches_done, nrow=5, normalize=True)
+    samples = state[tb.MODEL](batch, state)
+    save_image(samples, 'images/%d.png' % state[tb.BATCH], nrow=5, normalize=True)
 
 
 # Configure data loader
@@ -147,24 +151,25 @@ class d_loss(tb.metrics.Metric):
         return state[D_LOSS]
 
 
-from torchbearer.bases import base_closure
-
-
-closure_gen = base_closure(tb.X, tb.MODEL, tb.Y_PRED, tb.CRITERION, tb.LOSS, tb.OPTIMIZER)
+closure_gen = base_closure(tb.X, tb.MODEL, tb.Y_PRED, tb.CRITERION, tb.LOSS, GEN_OPT)
 closure_disc = base_closure(tb.Y_PRED, DISC_MODEL, DISC_IMGS, DISC_CRIT, tb.LOSS, DISC_OPT)
+
 
 def closure(self, state):
     closure_gen(self, state)
+    state[GEN_OPT].step()
     closure_disc(self, state)
     state[DISC_OPT].step()
 
 
-trial = tb.Trial(generator, optimizer_G, criterion=gen_crit, metrics=['loss', g_loss(), d_loss()],
+trial = tb.Trial(generator, None, criterion=gen_crit, metrics=['loss', g_loss(), d_loss()],
                             callbacks=[saver_callback], pass_state=True)
-trial.with_train_generator(dataloader)
+trial.with_train_generator(dataloader, steps=200000)
 trial.state[DISC_MODEL] = discriminator.to(device)
 trial.state[DISC_OPT] = optimizer_D
+trial.state[GEN_OPT] = optimizer_G
 trial.state[DISC_CRIT] = disc_crit
 trial.with_closure(closure)
 trial.to(device)
-trial.run(epochs=200)
+trial.run(epochs=1)
+