+++ tests/test_trial.py
@@ -1780,6 +1780,28 @@ class TestReplay(TestCase):
         self.assertTrue(callback.on_sample.call_count == 100)
         self.assertTrue(callback.on_sample_validation.call_count == 50)
 
+    def test_replay_none_train_steps(self):
+        t = Trial(MagicMock())
+        callback = MagicMock()
+        history = [((None, 5), {'test': i, 'val_test2': i+1}) for i in range(10)]
+
+        t.state[tb.HISTORY] = history
+        t.replay(callbacks=[callback], verbose=0)
+        self.assertEqual(callback.on_start.call_count, 1)
+        self.assertTrue(callback.on_sample.call_count == 0)
+        self.assertTrue(callback.on_sample_validation.call_count == 50)
+
+    def test_replay_none_validation_steps(self):
+        t = Trial(MagicMock())
+        callback = MagicMock()
+        history = [((10, None), {'test': i}) for i in range(10)]
+
+        t.state[tb.HISTORY] = history
+        t.replay(callbacks=[callback], verbose=0)
+        self.assertEqual(callback.on_start.call_count, 1)
+        self.assertTrue(callback.on_sample.call_count == 100)
+        self.assertTrue(callback.on_sample_validation.call_count == 0)
+
     def test_replay_one_batch_true(self):
         t = Trial(MagicMock())
         callback = MagicMock()
@@ -1851,17 +1873,20 @@ class TestTrialMembers(TestCase):
         torchmodel = "mod"
         optimizer = "opt"
         metric = tb.metrics.Metric('met')
+        cb = tb.callbacks.Callback()
+        cb.on_init = Mock()
 
-        torchbearertrial = Trial(torchmodel, optimizer, "crit", [metric], ["cb"])
-        correct_string = "--------------------- OPTIMZER ---------------------\nopt\n\n-------------------- CRITERION ---------------------\ncrit\n\n--------------------- METRICS ----------------------\n['met']\n\n-------------------- CALLBACKS ---------------------\n['cb']\n\n---------------------- MODEL -----------------------\nmod\n\n"
+        torchbearertrial = Trial(torchmodel, optimizer, "crit", [metric], [cb])
+        correct_string = "--------------------- OPTIMZER ---------------------\nopt\n\n-------------------- CRITERION ---------------------\ncrit\n\n--------------------- METRICS ----------------------\n['met']\n\n-------------------- CALLBACKS ---------------------\n['torchbearer.bases.Callback']\n\n---------------------- MODEL -----------------------\nmod\n\n"
         self.assertEqual(str(torchbearertrial), correct_string)
+        self.assertEqual(cb.on_init.call_count, 1)
 
     def test_repr(self):
         torchmodel = "mod"
         optimizer = "opt"
         metric = tb.metrics.Metric('met')
 
-        torchbearertrial = Trial(torchmodel, optimizer, "crit", [metric], ["cb"])
+        torchbearertrial = Trial(torchmodel, optimizer, "crit", [metric], [tb.callbacks.Callback()])
         self.assertEqual(str(torchbearertrial), repr(torchbearertrial))
 
     def test_train(self):