+++ torchbearer/metrics/primitives.py
@@ -19,6 +19,7 @@ def super(_, obj):
     return old_super(obj.__class__, obj)
 
 
+@default_for_key('binary_accuracy')
 @default_for_key('binary_acc')
 @running_mean
 @mean
@@ -47,9 +48,9 @@ class BinaryAccuracy(Metric):
         return torch.eq(y_pred, y_true).view(-1).float()
 
 
+@default_for_key('cat_accuracy')
 @default_for_key('cat_acc')
 @running_mean
-@std
 @mean
 class CategoricalAccuracy(Metric):
     """Categorical accuracy metric. Uses torch.max to determine predictions and compares to targets. Decorated with a
@@ -80,10 +81,11 @@ class CategoricalAccuracy(Metric):
         return (y_pred == y_true).float()
 
 
+@default_for_key('top_10_accuracy', k=10)
+@default_for_key('top_5_accuracy')
 @default_for_key('top_10_acc', k=10)
 @default_for_key('top_5_acc')
 @running_mean
-@std
 @mean
 class TopKCategoricalAccuracy(Metric):
     """Top K Categorical accuracy metric. Uses torch.topk to determine the top k predictions and compares to targets.
@@ -143,7 +145,6 @@ class MeanSquaredError(Metric):
 
 @default_for_key('loss')
 @running_mean
-@std
 @mean
 class Loss(Metric):
     """Simply returns the 'loss' value from the model state. Decorated with a mean, running_mean and std. Default for